add notification service, delete comment

+++ spring-project/src/main/java/org/example/springproject/ds/NotificationDto.java
@@ -9,8 +9,8 @@ import java.time.LocalDateTime;
 public class NotificationDto {
     private Long id;
 
-    private Long senderId; // The user who triggered the notification
-    private Long receiverId; // The user who receives the notification
+    private int senderId; // The user who triggered the notification
+    private int receiverId; // The user who receives the notification
     private String type; // e.g., "FRIEND_REQUEST", "FOLLOW", "BLOCK"
     private String message; // Optional: Custom message
     private boolean isRead = false; // Whether the notification has been read
